infrastructure/src/main/java/com/mercadolibre/planning/model/me/controller/WaveSuggestionController.java
package com.mercadolibre.planning.model.me.controller;

import com.mercadolibre.planning.model.me.controller.editor.WorkflowEditor;
import com.mercadolibre.planning.model.me.entities.projection.SimpleTable;
import com.mercadolibre.planning.model.me.gateways.planningmodel.dtos.Workflow;
import com.mercadolibre.planning.model.me.usecases.authorization.AuthorizeUser;
import com.mercadolibre.planning.model.me.usecases.authorization.dtos.AuthorizeUserDto;
import com.mercadolibre.planning.model.me.usecases.wavesuggestion.GetWaveSuggestion;
import com.mercadolibre.planning.model.me.usecases.wavesuggestion.dto.GetWaveSuggestionInputDto;
import com.newrelic.api.agent.Trace;
import lombok.AllArgsConstructor;
import org.springframework.beans.PropertyEditorRegistry;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.InitBinder;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;
import java.util.Optional;

import static com.mercadolibre.planning.model.me.gateways.authorization.dtos.UserPermission.OUTBOUND_PROJECTION;

@RestController
@AllArgsConstructor
@RequestMapping("/planning/model/middleend/warehouses/{warehouseId}/workflows/{workflow}")
public class WaveSuggestionController {

    private final AuthorizeUser authorizeUser;
    private final GetWaveSuggestion getWaveSuggestion;

    @Trace
    @GetMapping("/wave-suggestion")
    public ResponseEntity<SimpleTable> getWaveSuggestion(
            @PathVariable final String warehouseId,
            @PathVariable final Workflow workflow,
            @RequestParam("caller.id") final long callerId) {

        authorizeUser.execute(new AuthorizeUserDto(callerId, List.of(OUTBOUND_PROJECTION)));

        return ResponseEntity.of(Optional.of(getWaveSuggestion.execute(
                GetWaveSuggestionInputDto.builder()
                        .warehouseId(warehouseId)
                        .workflow(workflow)
                        .build()))
        );
    }

    @InitBinder
    public void initBinder(final PropertyEditorRegistry dataBinder) {
        dataBinder.registerCustomEditor(Workflow.class, new WorkflowEditor());
    }
}
